Rename: /service/server 파일 폴더화
